DAO: modification de la classe Group, de l'interface PersonDao et implements des methode de la classe PersonImpDao

+++ Projet1JEE/src/com/application/beans/Group.java
@@ -23,7 +23,7 @@ public class Group {
 	/**
 	 * @param idGroup
 	 */
-	public void setIdGroup(int idGroup) {
+	public void setIdGroup(long idGroup) {
 		this.idGroup = idGroup;
 	}
 	/**
